Stubbed out payments.
The GetPayment function is simple enough, so that's completely written,
but the rest is pretty complicated, so I just set up the foundation for
it and defined the method signatures.

The problem is that the difference between pending/charging/completed is
still unclear, and I'm not 100% sure how I want to handle that yet. I've
tried separating things out into distinct enough chunks that I have the
flexibility to actually charge these payments however I want in the
architectural scheme of things, but I'm having trouble deciding if the
API server itself should be making these charges, or if payments that
need to be charged should be put on a queue, charged by a worker, then
updated using the API. And when does a payment switch from pending to
charging? I think charging was meant to be a mutex to prevent
double-charging a payment by concurrent workers, but it doesn't make
sense in the other architecture. So I need to decide what that flow will
look like.