Implemented querying for payments.
In doing so, I introduced a horrible, horrible, no good dirty rotten
security vulnerability in the one table where security is probably the
most crucial. When querying for Payments, the status field allows for
SQL injection. Therefore, it's critical that the method only be used
with statuses that have been validated to match the pre-approved
statuses. This really needs to be fixed with a real SQL-generation
approach, but that will have to wait for now.